+++ packages/quick_actions/quick_actions_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.3/Dart 2.18.
+
+## 1.0.5
+
+* Fixes Java warnings.
+
 ## 1.0.4
 
 * Fixes compatibility with AGP versions older than 4.2.