+++ packages/web_benchmarks/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.3/Dart 2.18.
+
 ## 0.1.0+3
 
 * Migrates from SingletonFlutterWindow to PlatformDispatcher API.