+++ packages/webview_flutter/webview_flutter_wkwebview/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 3.4.3
+## 3.4.4
 
 * Adds support for `WebResouceError.failingUrl`.
+
+## 3.4.3
+
+* Replace `describeEnum` with the `name` getter.
 
 ## 3.4.2